Rename errors to err
