+++ client/base_client.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/pingcap/kvproto/pkg/pdpb"
 	"github.com/pingcap/log"
-	errs "github.com/pingcap/pd/v4/pkg/errors"
+	"github.com/pingcap/pd/v4/pkg/errs"
 	"github.com/pingcap/pd/v4/pkg/grpcutil"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
