+++ server/schedule/plugin_interface.go
@@ -19,7 +19,7 @@ import (
 	"sync"
 
 	"github.com/pingcap/log"
-	errs "github.com/pingcap/pd/v4/pkg/errors"
+	"github.com/pingcap/pd/v4/pkg/errs"
 	"go.uber.org/zap"
 )
