+++ server/systime_mon.go
@@ -18,7 +18,7 @@ import (
 	"time"
 
 	"github.com/pingcap/log"
-	errs "github.com/pingcap/pd/v4/pkg/errors"
+	"github.com/pingcap/pd/v4/pkg/errs"
 	"go.uber.org/zap"
 )
